### 2023-01-10 #5063 @editorialbot
[PRE REVIEW]: Openseize: A digital signal processing package for large EEG datasets in Python
**Submitting author:** @mscaudill (<a href="http://orcid.org/0000-0002-3656-9261">Matthew Caudill</a>)
**Repository:** https://github.com/mscaudill/openseize
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
