### 2023-01-12 #5063 @mscaudill
Hi @samhforbes, thank you for being the editor of my submission! Below is a list of possible reviewers selected based on their expertise in either, python software design, signal analysis, or computational neuroscience.

- jjerphan
- EtienneCmb
- sappelhoff
- cmaumet
