### 2023-01-17 e2dd0eb
docs: corrected DOI of woodman and cole papers
