+++ paper/paper.bib
@@ -68,7 +68,7 @@
   volume    = {72},
   number    = {8},
   pages     = {2031--2046},
-  doi       = {doi.org/10.3758/BF03196680},
+  doi       = {10.3758/BF03196680},
   publisher = {Springer},
 }
 
@@ -132,7 +132,7 @@
   volume    = {4},
   number    = {36},
   pages     = {1272},
-  doi       = {https://doi.org/10.21105/joss.01272},
+  doi       = {10.21105/joss.01272},
   publisher = {The Open Journal},
 }
 
