### 2023-01-19 #5063 @jjerphan
Hi @samhforbes, thank you for proposing. I do not have time for reviewing submission to JOSS currently but I may in a few months.
