### 2023-01-20 #5063 @lukassnoek
Hi Sam, I'm not very knowledgeable about EEG (and about to go on paternity leave), so perhaps someone else would be better!
