### 2023-02-27 d12b638
feat: additional path_utils that users will likely want in openseize
