+++ src/openseize/file_io/path_utils.py
@@ -1,0 +1,190 @@
+""""A collection of functions for working with path instances in Openseize.
+
+This module includes the following utilities:
+
+    - match: Regex matching of paths from two sequences of paths.
+    - mismatched: Locates paths from two sequences with mismatched path stems.
+    - rename: In-place path renaming by substring search & replacement.
+    - metadata: Path instance metadata extraction to a dictionary.
+"""
+
+import os
+from pathlib import Path
+from pathlib import PosixPath
+import re
+from typing import List, Pattern, Sequence, Set, Tuple, Union
+
+
+def match(files: Sequence[Path], others: Sequence[Path], pattern: Pattern 
+)-> (List[Tuple[Path, Path]]):
+    """Matches 2 equal lengthed sequences of Path instances using a regex
+    pattern common to both.
+
+    Args:
+        files:
+            A sequence of path instances.
+        others:
+            Another sequence of path instances.
+        pattern:
+            A regular expression pattern to match files with others by.
+
+    Returns:
+        A list of matched file tuples.
+
+    Raises:
+        ValueError if length of files does not match length of others or if
+        pattern is missing in any of files or others.
+        
+    Examples:
+        >>> files = [Path(x) for x in ['test_01_a.edf', 'test_02_b.edf']]
+        >>> others = [Path(x) for x in ['test_01.txt', 'test_02.txt']]
+        >>> # match from the start looking for letters then '_' then nums
+        >>> match(files, others, '^\w+_\d+') # doctest: +NORMALIZE_WHITESPACE
+        [(PosixPath('test_01_a.edf'), PosixPath('test_01.txt')), 
+        (PosixPath('test_02_b.edf'), PosixPath('test_02.txt'))]
+    """
+
+    # validate the number of passed files match
+    if len(files) != len(others):
+        msg = 'Number of paths in files and others must match: {} != {}'
+        raise ValueError(msg.format(len(files), len(others)))
+
+    # validate that the pattern exist in all passed files
+    if not all([re.search(pattern, str(x)) for x in files + others]):
+        msg = 'Pattern {} is missing from at least one file in files or others.'
+        raise ValueError(msg.format(pattern))
+
+    result = []
+    for fname in files:
+        matched = re.search(pattern, str(fname))
+        oname = [other for other in others if matched.group() in str(other)][0]
+        result.append((fname, oname))
+
+    return result
+
+
+def mismatched(files: Sequence[Path], others: Sequence[Path]) -> Set[Path]:
+    """Identifies mismatched Path stems between files and others.
+
+    Args:
+        files:
+            A sequence of Path instances.
+        others:
+            Another sequence of Path instances.
+    
+    Returns:
+        A set of stems from files that have no match in others.
+
+    Examples:
+        >>> files = [Path(x) for x in ['test_01.edf', 'test_02_b.edf']]
+        >>> others = [Path(x) for x in ['test_01.txt', 'test_02.text']]
+        >>> # find mismatches
+        >>> sorted(list(mismatched(files, others)))
+        ['test_02', 'test_02_b']
+    """
+
+    file_stems = set(fp.stem for fp in files)
+    other_stems = set(op.stem for op in others)
+    return file_stems.symmetric_difference(other_stems)
+
+
+def rename(files: Sequence[Path], substring: str, replacement: str):
+    """Renames each file in files by replacing substring in file name with
+    replacement.
+
+    Args:
+        files: 
+            A sequence of Path instances.
+        substr:
+            A substring in each file to be replaced.
+        replacement:
[…]
