### 2023-02-28 3b45518
feat: new path utils for simplifying the matching of files and metadata extraction
