+++ src/openseize/file_io/path_utils.py
@@ -2,108 +2,120 @@
 
 This module includes the following utilities:
 
-    - match: Regex matching of paths from two sequences of paths.
+    - re_match: Regex matching of paths from two sequences of paths.
     - mismatched: Locates paths from two sequences with mismatched path stems.
     - rename: In-place path renaming by substring search & replacement.
     - metadata: Path instance metadata extraction to a dictionary.
 """
 
-import os
 from pathlib import Path
-from pathlib import PosixPath
 import re
-from typing import List, Pattern, Sequence, Set, Tuple, Union
+from typing import Dict, List, Sequence, Set, Tuple, Union
 
 
-def match(files: Sequence[Path], others: Sequence[Path], pattern: Pattern 
+def re_match(paths: List[Path], others: List[Path], pattern: str
 )-> (List[Tuple[Path, Path]]):
-    """Matches 2 equal lengthed sequences of Path instances using a regex
-    pattern common to both.
+    r"""Matches 2 equal lengthed sequences of Path instances using a regex
+    pattern string common to both.
 
     Args:
-        files:
-            A sequence of path instances.
+        paths:
+            A list of path instances.
         others:
-            Another sequence of path instances.
+            Another list of path instances.
         pattern:
-            A regular expression pattern to match files with others by.
+            A regular expression pattern to match paths with others by.
 
     Returns:
-        A list of matched file tuples.
+        A list of matched path instance tuples.
 
     Raises:
-        ValueError if length of files does not match length of others or if
-        pattern is missing in any of files or others.
-        
+        ValueError if length of paths does not equal length of others or if
+        pattern is missing in any of paths or others or if a match can not be
+        found.
+
     Examples:
-        >>> files = [Path(x) for x in ['test_01_a.edf', 'test_02_b.edf']]
+        >>> paths = [Path(x) for x in ['test_01_a.edf', 'test_02_b.edf']]
         >>> others = [Path(x) for x in ['test_01.txt', 'test_02.txt']]
         >>> # match from the start looking for letters then '_' then nums
-        >>> match(files, others, '^\w+_\d+') # doctest: +NORMALIZE_WHITESPACE
-        [(PosixPath('test_01_a.edf'), PosixPath('test_01.txt')), 
+        >>> re_match(paths, others, '^\w+_\d+') # doctest: +NORMALIZE_WHITESPACE
+        [(PosixPath('test_01_a.edf'), PosixPath('test_01.txt')),
         (PosixPath('test_02_b.edf'), PosixPath('test_02.txt'))]
     """
 
-    # validate the number of passed files match
-    if len(files) != len(others):
-        msg = 'Number of paths in files and others must match: {} != {}'
-        raise ValueError(msg.format(len(files), len(others)))
+    # validate the number of passed paths match
+    if len(paths) != len(others):
+        msg = 'Number of paths in paths and others must match: {} != {}'
+        raise ValueError(msg.format(len(paths), len(others)))
 
-    # validate that the pattern exist in all passed files
-    if not all([re.search(pattern, str(x)) for x in files + others]):
-        msg = 'Pattern {} is missing from at least one file in files or others.'
-        raise ValueError(msg.format(pattern))
+    # validate that the pattern exist in all passed paths
+    missing = [str(fp) for fp in paths + others
+               if not re.search(pattern, str(fp))]
+    if missing:
+        msg = 'Pattern {} is missing in paths: {}'
+        raise ValueError(msg.format(pattern, missing))
 
     result = []
-    for fname in files:
-        matched = re.search(pattern, str(fname))
-        oname = [other for other in others if matched.group() in str(other)][0]
-        result.append((fname, oname))
+    for apath in paths:
[…]
