### 2023-03-01 4ef0dc0
docs: ensure examples use raw strings in regex
