### 2023-03-01 b0bdd79
docs: minor spacing issue fix in annotations for PEP8
