### 2023-03-02 304a62b
bug: path_utils re_match now looks only at the stems of paths to make matches and not full paths, test confirm working.
