### 2023-03-02 5c15754
feat: matching dialog now accepts a dir in addition to tkinter dialog names. dialogs.py now using type annotations and chks with mypy and pylint successful.
