+++ src/openseize/file_io/dialogs.py
@@ -1,16 +1,35 @@
+"""A collection of dialogs and message boxes.
+
+This module contains the following functions
+```
+    standard:
+        A unified dialog that can open any tkinter dialog by name and gracefully
+        destroys itself on successful exit or can be manually closed on error.
+
+    message:
+        A unified message box that can open any tkinter messagebox by name and
+        gracefully destroys itself on successful exit or can be manually closed
+        on error.
+
+    matching:
+        A dialog that uses regex patterns to match file stems of two sets of
+        file paths selected by path selection dialogs, or a directory dialog or
+        by manually specifying a directory containing all paths to match.
+```
+"""
+
+from collections import defaultdict
+from pathlib import Path
 import tkinter as tk
 import tkinter.filedialog as tkdialogs
 import tkinter.messagebox as tkmsgbox
-import re
-from pathlib import Path
-from collections import defaultdict
+from typing import List, Optional, Tuple, Union
+
 from openseize.file_io.path_utils import re_match
-from typing import List, Tuple
 
 
 def root_deco(dialog):
-    """Decorates a dialog with a toplevel that is destroyed when the dialog
-    is closed."""
+    """Decorates a dialog with a toplevel that is destroyed on dialog exit."""
 
     def decorated(*args, **kwargs):
         #create root and withdraw from screen
@@ -24,291 +43,122 @@ def root_deco(dialog):
 
 
 @root_deco
-def standard(kind, **options):
-    """Opens a tkinter modal file dialog & returns a Path instance.
-
-    Args:
-        kind (str):             name of a tkinter dialog
-    **options:
-        parent (widget):        ignored
-        title (str):            title of the dialog window
-        initialdir (str):       dir dialog starts in 
-        initialfile (str):      file selected on dialog open
-        filetypes (seq):        sequence of (label, pattern tuples) '*'
-                                wildcard allowed
-        defaultextension (str): default ext to append during save dialogs
-        multiple (bool):        when True multiple selection enabled
-    """
-
-    fp = getattr(tkdialogs, kind)(**options)
-    return Path(fp) if isinstance(fp, str) else [Path(p) for p in fp]
-
-
-@root_deco
-def message(kind, **options):
-    """Opens a tkinter modal message box & returns a string response.
-    
-    Args:
-        kind (str):         name of tkinter message box
-    **options:
-        default:            default button, if unspecified then 1st button
-        parent (widget):    ignored
-    """
-
-    return getattr(tkmsgbox, kind)(**options)
-
-
-def matching(kind: str, pattern: str, **options) -> List[Tuple[Path, Path]]:
-    r"""A dialog that regex pattern matches files with different suffixes
-    residing in a single directory or are selected with 2 file dialogs.
+def standard(kind: str, **options: str) -> Union[Path, List[Path]]:
+    """Opens a tkinter modal file dialog & returns a Path instance or list of
+    Path instances.
 
     Args:
         kind:
-            A tkinter dialog string name; must be 'askdirectory' or
-            'askopenfilenames'. If askdirectory, the files to be matched are
-            assumed to be in the selected directory. If askopenfilenames, the
-            user will use two dialogs to select the files to match.
+            Name of a tkinter file dialog.
+        options:
+            title (str):            title of the dialog window
+            initialdir (str):       dir dialog starts in
+            initialfile (str):      file selected on dialog open
+            filetypes (seq):        sequence of (label, pattern tuples) '*'
+                                    with wildcard allowed
+            defaultextension (str): default ext to append during save dialogs
+            multiple (bool):        when True multiple selection enabled
+
+    Returns:
+        A Path instance or list of Path instances.
+    """
+
+    # disable parent placement since root_deco automates this
[…]
