### 2023-03-02 2bbcef6
docs: adding the new dialogs into Openseize's documentation
