+++ README.md
@@ -7,8 +7,10 @@
   <i><font color='gray'>Digital Signal Processing for Big EEGs</font></i>
 </h2>
 
-![PyPI - License](https://img.shields.io/pypi/l/openseize?color=green&style=flat-square)
-![PyPI](https://img.shields.io/pypi/v/openseize?logo=pypi)
+![PyPI - License](https://img.shields.io/pypi/l/openseize?color=229954)
+![PyPI](https://img.shields.io/pypi/v/openseize?color=%238E44AD)
+![PyPI - Python Version](https://img.shields.io/pypi/pyversions/openseize)
+
 
 <p align="center"  style="font-size: 20px">
 <a href="#Key-Features">Key Features</a>   |  
