+++ README.md
@@ -8,42 +8,34 @@
 </h2>
 
 <p align="center">
-
   <a href="https://joss.theoj.org/papers/f737f4eb377a3bed6602dac51f6b13b4"><img 
     src="https://joss.theoj.org/papers/f737f4eb377a3bed6602dac51f6b13b4/status.svg" 
     alt="JOSS Review" />
   </a>
-
   <a href="https://github.com/mscaudill/openseize/blob/master/LICENSE"><img
     src="https://img.shields.io/pypi/l/openseize?color=teal" 
     alt="Openseize is released under the BSD 3-Clause license." />
   </a>
-  
   <a href="https://pypi.org/project/openseize/"><img 
     src="https://img.shields.io/pypi/v/openseize?color=%238E44AD)" 
     alt="Openseize pypi release" />
   </a>
-
   <a href="https://github.com/mscaudill/openseize/tree/master#Dependencies"><img 
     src="https://img.shields.io/pypi/pyversions/openseize" 
     alt="Python versions supported." />
   </a>
-
   <a href=""><img 
     src="https://img.shields.io/github/actions/workflow/status/mscaudill/openseize/test.yml" 
     alt="Openseize's build status" />
   </a>
-
- <a href=""><img 
+ <a href="https://github.com/mscaudill/openseize/pulls"><img 
     src="https://img.shields.io/badge/PRs-welcome-pink"
     alt="Pull Request Welcomed!" />
   </a>
-
  <a href=""><img 
     src="" 
     alt="" />
   </a>
-
 </p>
 
 
