### 2023-03-03 24878bb
docs: adjust pypi vers badge
