### 2023-03-03 c49ee5c
docs: adjust badge colors
