### 2023-03-04 72f0400
feat: removing old paths.py in openseize.tools since this is replaced with path_utils.py in file_io
