### 2023-03-04 269b98b
bugfix: fix spelling mistakes of openseize.io landing page
