### 2023-03-04 975dae2
bugfix: correct spelling errors in tests
