+++ src/openseize/__init__.py
@@ -1,3 +1,3 @@
 from .core.producer import producer
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"
