### 2023-03-04 1a13fe6
build: adding codespell to development dependencies
