+++ pyproject.toml
@@ -43,7 +43,8 @@ dependencies = [
 dev = ["isort",
     "mypy", 
     "pylint", 
-    "bumpver", 
+    "bumpver",
+    "codespell", 
     "mkdocs",
     "mkdocs-material",
     "mkdocstrings[python]",
