### 2023-03-10 mscaudill/openseize#3 @AJQuinn
[Quest]: JOSS Review
### Has this question been asked before?

- [X] I have searched the existing issues for this question.

### Question

Hi @mscaudill - apologies for the relatively slow review, I hope this wasn't too frustrating.

Short review - I'm very impressed! A lot of my usual queries and suggestions are already covered in OpenSeize - up to and including the spell checking that you updated recently. The documentation is first rate and I'm impressed by the code. It would be easy for something like your producer classes to become bloated and complicated but everything here is readable and straightforward (or as straightforward as something as involved as this could be...).

I have a couple of minor comments though it is up to you what you do with them. I can see two things that might slow down uptake of OpenSeize.

First is a reliance on EDF format. I suspect that a lot of potential users will stop looking at OpenSeize because of this and that would be a shame. I'd recommend at least adding in a roadmap for which other data formats might be supported and when (if they are) as well as some broad guidelines for how users of other data formats might use the package - eg converting data using another package or some additional information on how they can implement their own reader. Some of this maybe in the docs already but I'd recommend making it really prominant.

Second thought is that the extremely thorough and detailed documentation might be a bit intimidating for some users. The detail is fantastic - particularly on filtering - but makes it a bit tricky to see the big picture. I'd recommend including a couple of example 'pipeline' scripts to support the documentation. These could be a handful of short examples of common workflows using OpenSeize - eg taking a dataset -> low pass filtering -> downsampling -> notch filtering -> extracting a segment of interest -> writing a 'cleaned' file or perhaps taking a dataset -> downsampling -> computing PSDs for two time periods and visualising the difference. I think these would be great 'quick start' additions for less computational users.

I'd strongly encourange some changes as above but will leave the scope/scale of them up to you - let me know what you think.

Congratulations on the package - great work.

https://github.com/openjournals/joss-reviews/issues/5126

### Code of Conduct

- [X] I agree to follow this project's Code of Conduct
