### 2023-03-13 mscaudill/openseize#3 @mscaudill
Hi @AJQuinn,

Thank you for reviewing Openseize. I agree with both of the points you raised about the documentation. To make sure I fully address the first point, I am writing a small guide for users and developers who do not use the EDF format. I should finish this and expand the quickstart in the next few days.

Thanks,
Matt
