### 2023-03-14 da1ddcb
docs: modify edf_reader page docs to show the base class documentation
