### 2023-03-14 6189033
docs: new tutorial to guide users and developers wishing to use openseize with non-edf formats
