### 2023-03-16 1e8ada8
docs: show Reader ABC source code only without members
