### 2023-03-16 mscaudill/openseize#3 @mscaudill
Hi @AJQuinn

First, I want to express my thanks for suggesting specific improvements to Openseize's documentation. I think your ideas will make it easier for users and developers to learn and extend the software.

To that end, I added a new pipeline to the [quickstart guide](https://mscaudill.github.io/openseize/quickstart/) that walks newcomers through the process of making a sophisticated pipeline. This suggestion was great because it shows how Openseize can be used to construct arbitrarily complex DSP pipelines that are fully iterative.

Second, there is a new tutorial called [Other Formats](https://mscaudill.github.io/openseize/tutorials/other_formats/). This tutorial covers what to do when a user's EEG data is not in the EDF format. I present three strategies for handling non-EDF data with the highlight being the development of new Reader types. Specifically, I walk developers through constructing a prototype HDF5 reader. This demonstrates the set of methods developers must implement in their concrete readers to allow Openseize to produce data from their specific file types. Lastly, for non-computational users, I provide a roadmap for new reader development and encourage users to communicate in the github issues concerning concrete readers they would like maintainers and/or contributors to develop.

Best
Matt
