### 2023-03-17 mscaudill/openseize#3 @AJQuinn
Hi @mscaudill - the changes look great, thanks for getting behind them. I'll complete my part of the review.

Cheers,
Andrew
