### 2023-03-20 2550783
feat: producer serialization now supported by initializing reader producers with closed reader instances that are only temporarily opened during iteration. Multiprocessing testing examples appear to work. Clean-up and test to follow
