### 2023-03-21 4d7bb7a
feat: comments and docs for serialization added
