### 2023-03-22 #5126 @mscaudill
Hi @samhforbes

I've created a current release of the software and archived it to Zenodo. I also confirmed the authors, the ORIDs, and the license are correct and match between the paper and the archive.

The release version is [v1.1.0-joss](https://github.com/mscaudill/openseize/releases/tag/v1.1.0-joss)

The Zenodo DOI is [10.5281/zenodo.7760375](https://doi.org/10.5281/zenodo.7760375)  **Please note:** this is the DOI that cites all versions of the software and resolves to the latest version. If you need the DOI specific to the v1.1.0-joss release, that DOI is [10.5281/zenodo.7760376](https://doi.org/10.5281/zenodo.7760376)

Best
Matt
