### 2023-03-23 566eca3
test: remove dialog for data using the demo_data fixture in its place to avoid opening tkinter dialog in github runner.
