+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "openseize"
-version = "1.1.0"
+version = "1.2.0"
 authors =[{ name="Matthew Caudill", email="mscaudill@gmail.com" }]
 description = "Digital Signal Processing for Big EEG Datasets"
 readme = "README.md"
@@ -62,7 +62,7 @@ Homepage = "https://github.com/mscaudill/openseize"
 
 # bumpver configuration
 [tool.bumpver]
-current_version = "1.1.0"
+current_version = "1.2.0"
 version_pattern = "MAJOR.MINOR.PATCH[PYTAGNUM]"
 commit_message = 'bump version {old_version} -> {new_version}'
 commit = true
