### 2023-03-23 8152213
build: removing wget during conda install from env file triggering pypi version to install when openseize is pip installed
