### 2023-04-03 #5126 @mscaudill
Hi @samhforbes

I apologize for reaching back out to you since you turned Openseize over to the bcm-eic but I haven't heard anything from them in over a week and I wanted to make sure that I haven't been forgotten.

Thanks
Matt
