+++ CITATION.cff
@@ -1,0 +1,28 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Caudill
+  given-names: Matthew S.
+  orcid: "https://orcid.org/0000-0002-3656-9261"
+doi: 10.5281/zenodo.7760376
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Caudill
+    given-names: Matthew S.
+    orcid: "https://orcid.org/0000-0002-3656-9261"
+  date-published: 2023-04-05
+  doi: 10.21105/joss.05126
+  issn: 2475-9066
+  issue: 84
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 5126
+  title: "Openseize: A digital signal processing package for large EEG
+    datasets in Python"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.05126"
+  volume: 8
+title: "Openseize: A digital signal processing package for large EEG
+  datasets in Python"
